requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
Django==3.1.4
djangorestframework==3.12.2
install==1.3.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
Pygments==2.7.3
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
